Move colors to getColors
